requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.9.3
keras==2.9.0
pillow==10.1.0
numpy==1.23.5
python-multipart==0.0.6
jinja2==3.1.2
gunicorn==21.2.0
scikit-learn==1.2.2
